requirements.txt
Django==4.1.13
Pillow==10.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
whitenoise==6.5.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
djongo==1.3.6
pymongo==3.12.3 